Move Giveaway commands to TradeAdditionsModule.cs
